third_party/WebKit/LayoutTests/sensor/idl-Magnetometer.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

test(function() {
  // Test that Magnetometer interface exists
  assert_true('Magnetometer' in window);
}, 'Test that the Magnetometer interface is present.');

</script>
